+++ os/drivers/testcase/Make.defs
@@ -20,7 +20,16 @@
 ifeq ($(CONFIG_KERNEL_TEST_DRV),y)
 
 CSRCS += kernel_testcase_drv.c
-CSRCS += test_task_init.c
+CSRCS += test_task.c
+CSRCS += test_sched.c
+CSRCS += test_clock.c
+CSRCS += test_timer.c
+CSRCS += test_sem.c
+CSRCS += test_group.c
+
+ifneq ($(CONFIG_DISABLE_SIGNALS),y)
+CSRCS += test_signal.c
+endif
 
 CFLAGS += -I$(TOPDIR)/kernel
